requirements.txt
# ===== 1단계: API 호출 + S3 저장 필수 라이브러리 =====

# API 호출
requests==2.31.0            # 기상청 API HTTP 요청용 (가장 기본적인 HTTP 라이브러리)
tenacity==8.2.3            # API 재시도/백오프

# 데이터 처리
pandas==2.0.0              # API 응답 데이터를 테이블 형태로 처리 (DataFrame) - 현재 사용 중
numpy==1.24.0              # pandas 의존성, 수치 연산 (자동 설치되지만 명시) - 현재 사용 중

# 파일 포맷
pyarrow==12.0.0            # Parquet 파일 포맷 지원 (압축률 좋고 S3에 최적) - 현재 사용 중
s3fs==2024.2.0             # pandas/duckdb → s3:// 직접 읽기/쓰기

# AWS S3 연동
boto3==1.34.25             # AWS SDK - S3 업로드/다운로드
botocore==1.34.25          # boto3 의존성 (자동 설치되지만 버전 고정)

# 환경 설정
python-dotenv==1.0.1       # .env 파일에서 API키, AWS 자격증명 읽기
pydantic==1.10.11           # Airflow 2.6.3 호환 버전 (현재 사용 중)

# 설정 파일
PyYAML==6.0.1              # config.yaml 파일 파싱 (스테이션ID, 설정값 관리)

# 로깅
loguru==0.7.2              # 깔끔한 로그 출력 (print 대신 사용, 파일 로깅 가능)


